+++ dev/tasks/utils/tools.js
@@ -96,14 +96,15 @@ module.exports = {
 	/**
 	 * Gets the list of files that are supposed to be included in the next Git commit.
 	 *
+	 * @param cachedOnly {Boolean} List only files which are staged.
 	 * @returns {String[]} A list of file paths.
 	 */
-	getGitDirtyFiles: function() {
+	getGitDirtyFiles: function( cachedOnly ) {
 		// Cache it, so it is executed only once when running multiple tasks.
 		if ( !dirtyFiles ) {
 			dirtyFiles = this
 				// Compare the state of index with HEAD.
-				.shExec( 'git diff-index --name-only HEAD' )
+				.shExec( 'git diff-index --name-only ' + ( cachedOnly ? '--cached HEAD' : 'HEAD' ) )
 				// Remove trailing /n to avoid an empty entry.
 				.replace( /\s*$/, '' )
 				// Transform into array.
@@ -125,18 +126,15 @@ module.exports = {
 	 * @returns {String} The command output.
 	 */
 	shExec: function( command ) {
-		var sh = require( 'shelljs' );
-		sh.config.silent = true;
-
-		var ret = sh.exec( command );
+		var sh = require( 'shelljs' ),
+			ret = sh.exec( command, { silent: true } );
 
 		if ( ret.code ) {
 			throw new Error(
 				'Error while executing `' + command + '`:\n\n' +
-				ret.output
+				ret.stdout
 			);
 		}
-
-		return ret.output;
+		return ret.stdout;
 	}
 };