+++ gruntfile.js
@@ -18,7 +18,8 @@ module.exports = function( grunt ) {
 		'samples/toolbarconfigurator/lib/**',
 		'tests/adapters/jquery/_assets/**',
 		'tests/core/dom/_assets/**',
-		'tests/core/selection/_helpers/rangy.js'
+		'tests/core/selection/_helpers/rangy.js',
+		'tests/plugins/pastefromword/generated/_lib/q.js'
 	];
 
 	// Basic configuration which will be overloaded by the tasks.