+++ plugins/copyformatting/lang/it.js
@@ -1,0 +1,13 @@
+/*
+Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
+For licensing, see LICENSE.md or http://ckeditor.com/license
+*/
+CKEDITOR.plugins.setLang( 'copyformatting', 'it', {
+	label: 'Copia formattazione',
+	notification: {
+		copied: 'Formattazione copiata',
+		applied: 'Formattazione applicata',
+		canceled: 'Formattazione annullata',
+		failed: 'Formattazione fallita. Non è possibile applicare stili senza averli prima copiati.'
+	}
+} );