+++ plugins/link/plugin.js
@@ -92,9 +92,11 @@
 			CKEDITOR.dialog.add( 'anchor', this.path + 'dialogs/anchor.js' );
 
 			editor.on( 'doubleclick', function( evt ) {
-				var element = CKEDITOR.plugins.link.getSelectedLink( editor ) || evt.data.element;
+				// If the link has descendants and the last part of it is also a part of a word partially
+				// unlinked, clicked element may be a descendant of the link, not the link itself. (#11956)
+				var element = CKEDITOR.plugins.link.getSelectedLink( editor ) || evt.data.element.getAscendant( 'a', 1 );
 
-				if ( !element.isReadOnly() ) {
+				if ( element && !element.isReadOnly() ) {
 					if ( element.is( 'a' ) ) {
 						evt.data.dialog = ( element.getAttribute( 'name' ) && ( !element.getAttribute( 'href' ) || !element.getChildCount() ) ) ? 'anchor' : 'link';